agri_ai_agent/requirements.txt
streamlit
requests
google-generativeai
fastapi
pydantic
python-dotenv
uvicorn
